moved patched shader printing into its own file
