+++ src/main/java/net/coderbot/iris/postprocess/FinalPassRenderer.java
@@ -25,7 +25,7 @@ import net.coderbot.iris.gl.program.ProgramBuilder;
 import net.coderbot.iris.gl.program.ProgramSamplers;
 import net.coderbot.iris.gl.program.ProgramUniforms;
 import net.coderbot.iris.gl.sampler.SamplerLimits;
-import net.coderbot.iris.pipeline.DeferredWorldRenderingPipeline;
+import net.coderbot.iris.pipeline.*;
 import net.coderbot.iris.pipeline.transform.TransformPatcher;
 import net.coderbot.iris.rendertarget.Blaze3dRenderTargetExt;
 import net.coderbot.iris.rendertarget.RenderTarget;
@@ -283,7 +283,7 @@ public class FinalPassRenderer {
 			String vertex = TransformPatcher.patchCompositeDepth(source.getVertexSource().orElse(null));
 			String geometry = TransformPatcher.patchCompositeDepth(source.getGeometrySource().orElse(null));
 			String fragment = TransformPatcher.patchCompositeDepth(source.getFragmentSource().orElse(null));
-			DeferredWorldRenderingPipeline.debugPatchedShaders(source.getName(), vertex, geometry, fragment);
+			PatchedShaderPrinter.debugPatchedShaders(source.getName(), vertex, geometry, fragment);
 			builder = ProgramBuilder.begin(source.getName(), vertex, geometry, fragment,
 				IrisSamplers.COMPOSITE_RESERVED_TEXTURE_UNITS);
 		} catch (RuntimeException e) {
